Basic framework working

+++ rrdebug.py
@@ -3,6 +3,9 @@
 from cmd import Cmd
 import subprocess
 import pexpect
+import os
+import logging
+import time
 
 class RrDebugger(Cmd):
     def __init__(self):
@@ -18,13 +21,21 @@ class RrDebugger(Cmd):
         self.image = None
         self.qemu_process = None
         self.gdb_pexpect = None
+        self.qemu_cwd = None
 
     def read_init_file(self, filename):
         f = open(filename, 'r')
         for line in f:
             self.onecmd(line)
+        
+        logging.info('Init file loaded.')
 
     def do_EOF(self, line):
+        '''Kill qemu and gdb and exit.'''
+        if (self.gdb_pexpect is not None):
+            self.gdb_pexpect.kill(0)
+        if (self.qemu_process is not None):
+            self.qemu_process.kill()
         return True
 
     def do_set_vmlinux(self, line):
@@ -42,9 +53,13 @@ class RrDebugger(Cmd):
         if (len(line_split) > 1):
             self.qemu_args.extend(line_split[1:])
 
+    def do_set_qemu_cwd(self, line):
+        '''Specify cwd for qemu.'''
+        self.qemu_cwd = line
+
     def do_set_replay_file(self, line):
         '''Specify the replay file.'''
-        self.qemu_replay_file = 'line'
+        self.qemu_replay_file = line
 
     def do_set_image(self, line):
         '''Specify the qcow2 image file.'''
@@ -52,36 +67,53 @@ class RrDebugger(Cmd):
 
     def do_start(self, line):
         '''Start qemu and attach gdb to it.'''
-        if (self.qemu_process.poll() is None):
+        if (self.qemu_process is not None and self.qemu_process.poll() is None):
             self.qemu_process.kill()
 
-        self.qemu_process = subprocess.Popen([qemu_exec].extend(qemu_args).\
-                extend([qemu_replay_keyword, qemu_replay_file, image]),
+        cmd_line = [self.qemu_exec]
+        cmd_line.extend(self.qemu_args)
+        cmd_line.extend([self.qemu_replay_keyword, self.qemu_replay_file,
+            self.image])
+
+        logging.debug("qemu cmd_line = " + " ".join(cmd_line))
+        self.qemu_process = subprocess.Popen(cmd_line, cwd=self.qemu_cwd,
                 stdout = subprocess.PIPE, stdin = subprocess.PIPE,
                 stderr = subprocess.PIPE)
+
+        time.sleep(5)
+        #logging.debug(self.qemu_process.stdout.read())
         
         if (self.gdb_pexpect is None):
             self.gdb_pexpect = pexpect.spawn(self.gdb_exec)
-            self.gdb_pexpect.expect('(gdb)')
+            self.gdb_pexpect.expect('\(gdb\)')
 
-        self.gdb_pexpect.sendline(self.gbd_connect_cmd)
-        self.gdb_pexpect.expect('(gdb)')
+        #self.gdb_pexpect.interact()
+        self.gdb_pexpect.sendline(self.gdb_connect_cmd)
+        self.gdb_pexpect.expect('\(gdb\)')
+
+        logging.debug(self.gdb_pexpect.before)
         
-        self.gdb_pexpect.sendline('file ' + self.vm_linux)
-        self.gdb_pexpect.expect('(gdb)')
+        #self.gdb_pexpect.interact()
+        self.gdb_pexpect.sendline('file ' + self.vmlinux)
+        self.gdb_pexpect.expect('y or n')
+        self.gdb_pexpect.sendline('y')
+        self.gdb_pexpect.expect('\(gdb\)')
+        logging.debug(self.gdb_pexpect.before)
 
     def do_watch(self, line):
         '''watch [var_name]
         Watch the specified kernel variable.'''
 
         self.gdb_pexpect.sendline('watch ' + line)
-        self.gdb_pexpect.expect('(gdb)')
+        self.gdb_pexpect.expect('\(gdb\)')
+        #logging.debug(self.gdb_expect.before)
 
     def do_run_till_live(self, line):
         pass
 
 
 if __name__ == '__main__':
+    logging.basicConfig(level=logging.DEBUG)
     dbg = RrDebugger()
     if os.path.exists('.rrdebuginit'):
         dbg.read_init_file('.rrdebuginit')
